requirements.txt
# Core Spotify and audio analysis
spotipy>=2.23.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
requests>=2.31.0

# Web application (for enhanced interface)
Flask>=2.3.3
Flask-SocketIO>=5.3.6

# Date utilities
python-dateutil>=2.8.2

# Audio analysis (optional but recommended)
librosa>=0.10.1
soundfile>=0.12.1
